% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_object.R
\name{plot.consortflow}
\alias{plot.consortflow}
\title{Plot a consortflow object}
\usage{
\method{plot}{consortflow}(x, ...)
}
\arguments{
\item{x}{A \code{consortflow} object.}

\item{...}{Additional arguments passed to \code{\link[=cf_draw]{cf_draw()}}.}
}
\description{
Draws the exclusion flowchart to the current graphics device.
Wrapper around \code{\link[=cf_draw]{cf_draw()}}.
}
