% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_object.R
\name{print.consortflow}
\alias{print.consortflow}
\title{Print a consortflow object}
\usage{
\method{print}{consortflow}(x, ...)
}
\arguments{
\item{x}{A \code{consortflow} object.}

\item{...}{Ignored.}
}
\description{
Print a consortflow object
}
